Save application when funding information supplied
We verify that task list status is 'Not started' at outset
and moves to Completed once the form contents are valid and
successfully submitted.

NB we need to re-stub the "get application" API call with a
`data` object containing the key of the _page_ which is now
complete:

```
data: {
  'area-and-funding': {
    'funding-information': {},
  },
```

See `form-pages/utils/getTaskStatus.ts` for details.