add utility functions for dates
Referrers need to be able to add to a list of ACCTs (Assessment, Care in
Custody and Teamwork). To do this we will have one page that renders the
list of ACCTs added, and one page with a form to add the data.

This adds the utility functions from CAS1 needed to add dates to a form.