add ACCT data page template
Referrers need to be able to add to a list of ACCTs (Assessment, Care in
Custody and Teamwork). To do this we will have one page that renders the
list of ACCTs added, and one page with a form to add the data.

This adds the template with that form.