add new endpoint to append to list on Application
Referrers need to be able to add to a list of ACCTs (Assessment, Care in
Custody and Teamwork). To do this we will have one page that renders the
list of ACCTs added, and one page with a form to add the data.

We have decided to use a custom end point here and a form outside of the
regular "Task List system". We will shortly add an ADR describing this
process, but the TLDR is that the ACCT data form will PUT to this new
endpoint to append to a list.